requirements.txt
# requirements.txt (updated)
Flask==2.3.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
flasgger==0.9.5
Flask-Limiter==3.5.0
Werkzeug==2.3.7
# For testing
pytest==7.4.0
pytest-flask==1.3.0
API_KEY

SECRET_KEY

LOG_LEVEL

FLASK_ENV